Saving neuron model using passthru neurons
Passthru is a neuron type that can send spike events weighted with the
neuron's current/voltage. It can be used to inject "image" values which
are not black & white. It can also be used to organize into groups the
input synapses to a neuron or delay some synapse (although not ideal).